added webapi changes for minimal hosting
