+++ src/Scaffolding/VS.Web.CG.EFCore/ICodeModelService.cs
@@ -1,8 +1,7 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-using Microsoft.DotNet.Scaffolding.Shared.Project;
 using System.Threading.Tasks;
+using Microsoft.DotNet.Scaffolding.Shared.Project;
 
 namespace Microsoft.VisualStudio.Web.CodeGeneration.EntityFrameworkCore
 {
