+++ src/Scaffolding/VS.Web.CG.EFCore/NewDbContextTemplateModel.cs
@@ -1,10 +1,10 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using Microsoft.DotNet.Scaffolding.Shared.Project;
 using System;
 using System.Collections.Generic;
-
+using Microsoft.DotNet.Scaffolding.Shared.Project;
+    
 namespace Microsoft.VisualStudio.Web.CodeGeneration.EntityFrameworkCore
 {
     public class NewDbContextTemplateModel
