+++ test/Shared/Microsoft.DotNet.Scaffolding.Shared.Tests/ProjectModifierHelperTests.cs
@@ -1,0 +1,103 @@
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using Microsoft.CodeAnalysis;
+using Microsoft.DotNet.Scaffolding.Shared.Project;
+using Moq;
+using Xunit;
+
+namespace Microsoft.DotNet.Scaffolding.Shared.Tests
+{
+    public class ProjectModifierHelperTests : DocumentBuilderTestBase
+    {
+        [Theory]
+        [InlineData(new object[] { new string[] { "Startup.cs", "File.cs", "Test", "", null},
+                                   new string[] { "Startup", "File", "", "", "" } })]
+        public void GetClassNameTests(string[] classNames, string[] formattedClassNames)
+        {
+            for (int i = 0; i < classNames.Length; i++)
+            {
+                string className = classNames[i];
+                string formattedClassName = formattedClassNames[i];
+                Assert.Equal(ProjectModifierHelper.GetClassName(className), formattedClassName);
+            }
+        }
+
+        [Fact]
+        public async Task GetStartupClassNameTests()
+        {
+            Document programDocument = CreateDocument(ProgramCsFile);
+            Document programDocumentNoStartup = CreateDocument(ProgramCsFileNoStartup);
+            Document programDocumentDifferentStartup = CreateDocument(ProgramCsFileWithDifferentStartup);
+
+            string startupName = await ProjectModifierHelper.GetStartupClassName(programDocument);
+            string emptyStartupName = await ProjectModifierHelper.GetStartupClassName(programDocumentNoStartup);
+            string notStartupName = await ProjectModifierHelper.GetStartupClassName(programDocumentDifferentStartup);
+            string nullStartup = await ProjectModifierHelper.GetStartupClassName(null);
+
+            Assert.Equal("Startup", startupName);
+            Assert.Equal("", emptyStartupName);
+            Assert.Equal("", nullStartup);
+            Assert.Equal("NotStartup", notStartupName);
+        }
+
+        [Theory]
+        [InlineData(new object[] { new string[] { "string.ToString()", "bool.ToString()", "IServices.AddRazorPages()", "IWebSomething.DoWebStuff()", "", "Wrong.Missing()" },
+                                   new string[] { "password.ToString()", "IsTrue.ToString()", "services.AddRazorPages()", "app.DoWebStuff()", "", "Wrong.Missing()" } })]
+        public void FormatCodeBlockTests(string[] parameterKeys, string[] formattedValues)
+        {
+            for (int i = 0; i < parameterKeys.Length; i++)
+            {
+                string paramKey = parameterKeys[i];
+                string correctCodeBlock = formattedValues[i];
+                string formattedCodeBlock = ProjectModifierHelper.FormatCodeBlock(paramKey, ParameterValues);
+                Assert.Equal(formattedCodeBlock, correctCodeBlock);
+            }
+        }
+
+
+        [Fact]
+        public async Task IsMinimalAppTests()
+        {
+            var modelTypes = new List<ModelType>();
+            var minimalModelTypesLocator = new Mock<IModelTypesLocator>();
+            var nonMinimalTypesLocator = new Mock<IModelTypesLocator>();
+            //IModelTypesLocator has a Startup.cs file
+            nonMinimalTypesLocator.Setup(m => m.GetType("Startup")).Returns(() => { return new List<ModelType> { startupModel }; });
+            //IModelTypesLocator does not have a Startup.cs file
+            minimalModelTypesLocator.Setup(m => m.GetType("Startup")).Returns(() => { return new List<ModelType> { }; });
+            Assert.True(await ProjectModifierHelper.IsMinimalApp(minimalModelTypesLocator.Object));
+            Assert.False(await ProjectModifierHelper.IsMinimalApp(nonMinimalTypesLocator.Object));
+        }
+
+        [Theory]
+        [InlineData(new object[] {
+                        new string[] { "\nWebApplication.CreateBuilder.Services.AddAuthentication(OpenIdConnectDefaults.AuthenticationScheme)\n    .AddMicrosoftIdentityWebApp(builder.Configuration.GetSection(\"AzureAd\"));",
+                                       "\nWebApplication.CreateBuilder.Services.AddAuthorization(options =>\n{\n    options.FallbackPolicy = options.DefaultPolicy;\n});\n\r",
+                                       "absdf \n \r asdfsadfasdf sdfasdf \n {} ",
+                                       "",
+                                       null },
+                        new string[] { "WebApplication.CreateBuilder.Services.AddAuthentication(OpenIdConnectDefaults.AuthenticationScheme).AddMicrosoftIdentityWebApp(builder.Configuration.GetSection(\"AzureAd\"))",
+                                       "WebApplication.CreateBuilder.Services.AddAuthorization(options=>{options.FallbackPolicy=options.DefaultPolicy})",
+                                       "absdfasdfsadfasdfsdfasdf{}",
+                                       "",
+                                       ""} }
+        )]
+        public void TrimStatementTests(string[] statements, string[] trimmedStatements)
+        {
+            for (int i = 0; i < statements.Length; i++)
+            {
+                string statement = statements[i];
+                string trimmedStatement = trimmedStatements[i];
+                Assert.Equal(trimmedStatement, ProjectModifierHelper.TrimStatement(statement));
+            }
+        }
+
+        private static readonly ModelType startupModel = new ModelType
+        {
+            Name = "Startup",
+            Namespace = "Application",
+            FullName = "C:\\Solution\\Application\\Startup.cs"
+        };
+    }
+}
